src/tools/medgemma_tool.py
"""MedGemma integration tool for medical image analysis."""
import logging
from typing import Optional, Dict, Any
import base64
from io import BytesIO
from PIL import Image

logger = logging.getLogger(__name__)


class MedGemmaTool:
    """Tool for interacting with MedGemma model."""

    def __init__(self, endpoint: str = "local", model_path: str = "google/medgemma-4b"):
        """
        Initialize MedGemma tool.

        Args:
            endpoint: Either 'local' or 'vertex_ai'
            model_path: Path to the MedGemma model
        """
        self.endpoint = endpoint
        self.model_path = model_path
        logger.info(f"Initializing MedGemma tool with endpoint: {endpoint}")

    def analyze_image(self, image_base64: str, prompt: Optional[str] = None) -> str:
        """
        Analyze a medical image using MedGemma.

        Args:
            image_base64: Base64 encoded image
            prompt: Optional analysis prompt

        Returns:
            Analysis results as a string
        """
        try:
            # Decode and validate image
            image_data = base64.b64decode(image_base64)
            image = Image.open(BytesIO(image_data))

            logger.info(f"Analyzing image of size: {image.size}, mode: {image.mode}")

            # For MVP/hackathon: Return mock analysis
            # In production, this would call actual MedGemma API
            if self.endpoint == "local":
                return self._mock_medgemma_analysis(image, prompt)
            else:
                return self._vertex_ai_analysis(image, prompt)

        except Exception as e:
            logger.error(f"Error analyzing image with MedGemma: {e}")
            return f"Error during analysis: {str(e)}"

    def _mock_medgemma_analysis(self, image: Image.Image, prompt: Optional[str]) -> str:
        """
        Mock MedGemma analysis for demo purposes.

        In a production environment, this would be replaced with actual
        MedGemma API calls via Hugging Face or Vertex AI.
        """
        analysis = """
Medical Image Analysis Results:

FINDINGS:
1. Chest X-Ray - Frontal View
   - Quality: Adequate penetration and positioning
   - Heart: Normal size and contour (Cardiothoracic ratio < 0.5)
   - Lungs: Clear lung fields bilaterally
   - Pleura: No pleural effusion or pneumothorax detected
   - Bones: No acute fractures visible

2. Possible Observations:
   - Mild linear opacity in right lower lung zone - likely subsegmental atelectasis
   - No focal consolidation
   - Vascular markings appear normal

IMPRESSION:
- Essentially normal chest radiograph
- Consider clinical correlation for the subtle right lower lung finding
- No acute cardiopulmonary abnormality identified

CONFIDENCE: 85%
        """

        if prompt:
            analysis = f"Analysis based on user prompt: '{prompt}'\n\n" + analysis

        return analysis.strip()

    def _vertex_ai_analysis(self, image: Image.Image, prompt: Optional[str]) -> str:
        """
        Placeholder for Vertex AI MedGemma integration.

        This would use Google Cloud's Vertex AI to call MedGemma.
        """
        logger.warning("Vertex AI endpoint not yet implemented, using mock data")
        return self._mock_medgemma_analysis(image, prompt)

    def get_tool_definition(self) -> Dict[str, Any]:
        """
        Return the function definition for Gemini Function Calling.

        This enables Gemini to automatically call this tool when needed.
        """
        return {
            "name": "analyze_medical_image",
            "description": (
                "Analyze a medical image (X-ray, CT, MRI) using the specialized "
                "MedGemma model. Returns detailed findings including anatomical "
                "observations, abnormalities, and diagnostic impressions."
            ),
            "parameters": {
                "type": "object",
                "properties": {
                    "image_base64": {
                        "type": "string",
                        "description": "Base64 encoded medical image"
                    },
                    "focus_areas": {
                        "type": "string",
                        "description": (
                            "Optional: Specific areas to focus on "
                            "(e.g., 'lung fields', 'cardiac silhouette')"
                        )
                    }
                },
                "required": ["image_base64"]
            }
        }
